task(Avatar): Component architecture improved
