+++ components/base/Avatar/components/Picture.tsx
@@ -5,13 +5,8 @@ import gradient from 'random-gradient';
 
 import Icon from 'components/ui/Icon';
 
-import {
-  AVATAR_VARIANT_BADGE,
-  AVATAR_VARIANT_BANNER,
-  AVATAR_VARIANT_EDIT,
-  AVATAR_VARIANT_MOSAIC,
-  AVATAR_VARIANT_TYPE,
-} from '../Avatar';
+import { AVATAR_VARIANT_TYPE } from '../interfaces';
+import { getPictureSize, getPictureFontSize } from '../utils';
 
 interface Props {
   className?: string;
@@ -24,36 +19,6 @@ interface Props {
   variant?: AVATAR_VARIANT_TYPE;
   walletId?: string;
 }
-
-const pictureSize = (variant?: AVATAR_VARIANT_TYPE) => {
-  switch (variant) {
-    case AVATAR_VARIANT_BANNER:
-      return '12rem';
-    case AVATAR_VARIANT_EDIT:
-      return '9.6rem';
-    case AVATAR_VARIANT_BADGE:
-      return '3.6rem';
-    case AVATAR_VARIANT_MOSAIC:
-      return '8rem';
-    default:
-      return '5.6rem';
-  }
-};
-
-const fontSize = (variant?: AVATAR_VARIANT_TYPE) => {
-  switch (variant) {
-    case AVATAR_VARIANT_BANNER:
-      return '5.6rem';
-    case AVATAR_VARIANT_EDIT:
-      return '3.2rem';
-    case AVATAR_VARIANT_BADGE:
-      return '2rem';
-    case AVATAR_VARIANT_MOSAIC:
-      return '2.8rem';
-    default:
-      return '2.4rem';
-  }
-};
 
 const Picture = ({
   className,
@@ -108,8 +73,8 @@ const SPictureContainer = styled.div<{
 `;
 
 const SPictureWrapper = styled.div<{ variant?: AVATAR_VARIANT_TYPE }>`
-  width: ${({ variant }) => pictureSize(variant)};
-  height: ${({ variant }) => pictureSize(variant)};
+  width: ${({ variant }) => getPictureSize(variant)};
+  height: ${({ variant }) => getPictureSize(variant)};
   position: relative;
   border-radius: 50%;
   box-shadow: 0 0.2rem 0.2rem rgba(0, 0, 0, 0.25);
@@ -160,7 +125,7 @@ const SInitials = styled.div<{ isClickable?: boolean; name: string }>`
 const SLetter = styled.div<{ variant?: AVATAR_VARIANT_TYPE }>`
   color: ${({ theme }) => theme.colors.invertedContrast};
   font-family: ${({ theme }) => theme.fonts.medium};
-  font-size: ${({ variant }) => fontSize(variant)};
+  font-size: ${({ variant }) => getPictureFontSize(variant)};
   text-transform: uppercase;
 `;
