+++ components/base/Avatar/constants.ts
@@ -1,0 +1,5 @@
+export const AVATAR_VARIANT_BADGE = 'badge';
+export const AVATAR_VARIANT_BANNER = 'banner';
+export const AVATAR_VARIANT_EDIT = 'edit';
+export const AVATAR_VARIANT_MOSAIC = 'mosaic';
+export const AVATAR_VARIANT_TRANSACTION = 'transaction';
