+++ components/base/Avatar/index.tsx
@@ -1,4 +1,5 @@
-export { default, AVATAR_VARIANT_BADGE, AVATAR_VARIANT_BANNER, AVATAR_VARIANT_EDIT, AVATAR_VARIANT_MOSAIC } from './Avatar';
-export type { AVATAR_VARIANT_TYPE } from './Avatar';
+export { default } from './Avatar';
 export { default as Banner } from './components/Banner';
 export { default as Picture } from './components/Picture';
+export * from './constants';
+export type { AVATAR_VARIANT_TYPE } from './interfaces';
