+++ components/base/Avatar/interfaces.ts
@@ -1,0 +1,14 @@
+import {
+  AVATAR_VARIANT_BADGE,
+  AVATAR_VARIANT_BANNER,
+  AVATAR_VARIANT_EDIT,
+  AVATAR_VARIANT_MOSAIC,
+  AVATAR_VARIANT_TRANSACTION,
+} from './constants';
+
+export type AVATAR_VARIANT_TYPE =
+  | typeof AVATAR_VARIANT_BADGE
+  | typeof AVATAR_VARIANT_BANNER
+  | typeof AVATAR_VARIANT_EDIT
+  | typeof AVATAR_VARIANT_MOSAIC
+  | typeof AVATAR_VARIANT_TRANSACTION;
